requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
async-generator==1.10
attrs==23.1.0
autoflake==2.1.1
Automat==22.10.0
backcall==0.2.0
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
cryptography==41.0.1
cssselect==1.2.0
dataclasses==0.6
decorator==5.1.1
Django==4.2.3
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.12.2
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipdb==0.13.13
ipython==8.14.0
isort==5.12.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.18.2
jmespath==1.0.1
lxml==4.9.2
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.1
parsel==1.8.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.8.0
prompt-toolkit==3.0.38
Protego==0.2.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyflakes==3.0.1
Pygments==2.15.1
pyOpenSSL==23.2.0
PySocks==1.7.1
PyYAML==6.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.9.0
scrapy-selenium==0.0.7
scrapy-user-agents==0.1.1
-e git+https://github.com/kekesh/sd.git@11d510202dbc187de5adceb323af80e889b3e2c1#egg=sd
selenium==4.9.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
stack-data==0.6.2
tldextract==3.4.4
tomli==2.0.1
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.3
Twisted==22.10.0
typing_extensions==4.6.3
ua-parser==0.16.1
urllib3==1.26.16
user-agents==2.2.0
w3lib==2.1.1
wcwidth==0.2.6
wsproto==1.2.0
zope.interface==6.0
